[SPR-93] feat: @CurrentUser 적용

+++ src/main/java/com/climeet/climeet_backend/domain/climbingrecord/ClimbingRecordService.java
@@ -12,6 +12,7 @@ import com.climeet.climeet_backend.domain.routerecord.RouteRecordRepository;
 import com.climeet.climeet_backend.domain.routerecord.RouteRecordService;
 import com.climeet.climeet_backend.domain.routerecord.dto.RouteRecordRequestDto.CreateRouteRecordDto;
 import com.climeet.climeet_backend.domain.routerecord.dto.RouteRecordResponseDto.RouteRecordSimpleInfo;
+import com.climeet.climeet_backend.domain.user.User;
 import com.climeet.climeet_backend.global.response.code.status.ErrorStatus;
 import com.climeet.climeet_backend.global.response.exception.GeneralException;
 import jakarta.persistence.Tuple;
@@ -23,8 +24,10 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.ErrorResponseException;
 
 @RequiredArgsConstructor
 @Service
@@ -42,27 +45,29 @@ public class ClimbingRecordService {
      * 클라이밍기록 생성(루트기록생성포함)
      */
     @Transactional
-    public ResponseEntity<String> createClimbingRecord(CreateClimbingRecordDto requestDto) {
+    public ResponseEntity<String> createClimbingRecord(User user, CreateClimbingRecordDto requestDto) {
         ClimbingGym climbingGym = gymRepository.findById(requestDto.getGymId())
             .orElseThrow(() -> new GeneralException(ErrorStatus._CLIMBING_RECORD_NOT_FOUND));
 
         ClimbingRecord savedClimbingRecord = climbingRecordRepository
-            .save(ClimbingRecord.toEntity(requestDto, climbingGym));
+            .save(ClimbingRecord.toEntity(user, requestDto, climbingGym));
 
         //선택받았으니 하나 추가
         climbingGym.thisWeekSelectionCountUp();
+        Long totalTime = requestDto.getTime().toNanoOfDay()/1000000000;
+        user.thisWeekTotalClimbingTimeUp(totalTime);
 
         List<CreateRouteRecordDto> routeRecords = requestDto.getRouteRecordRequestDtoList();
         // 루트기록 리퀘스트 돌면서 루트 리퀘스트 저장
 
         routeRecords.forEach(
-            routeRecord -> routeRecordService.addRouteRecord(routeRecord, savedClimbingRecord));
+            routeRecord -> routeRecordService.addRouteRecord(user, routeRecord, savedClimbingRecord));
 
         return ResponseEntity.ok("클라이밍 기록생성 성공");
     }
 
     /**
-     * 클라이밍 간편기록 전체 조회
+     * 클라이밍 간편기록 전체 조회 - ID 상관없이
      */
     public List<ClimbingRecordSimpleInfo> getClimbingRecords() {
         List<ClimbingRecord> recordList = climbingRecordRepository.findAll();
@@ -96,20 +101,19 @@ public class ClimbingRecordService {
     /**
      * 클라이밍 간편기록 날짜범위조회
      */
-    public List<ClimbingRecordSimpleInfo> getClimbingRecordsBetweenLocalDates(LocalDate startDate,
+    public List<ClimbingRecordSimpleInfo> getClimbingRecordsBetweenLocalDates(User user, LocalDate startDate,
         LocalDate endDate) {
         if (startDate.isAfter(endDate)) {
             throw new GeneralException(ErrorStatus._INVALID_DATE_RANGE);
         }
 
-        List<ClimbingRecord> climbingRecordList = climbingRecordRepository.findByClimbingDateBetween(
-            startDate, endDate);
+        List<ClimbingRecord> climbingRecordList = climbingRecordRepository.findByClimbingDateBetweenAndUser(
+            startDate, endDate, user);
 
         if(climbingRecordList.isEmpty()){
             throw new GeneralException(ErrorStatus._EMPTY_CLIMBING_RECORD);
         }
 
-        // Dto로 변환
         return climbingRecordList.stream()
             .map(record -> ClimbingRecordSimpleInfo.toDTO(record))
             .collect(Collectors.toList());
@@ -121,11 +125,15 @@ public class ClimbingRecordService {
      */
     // TODO: 2024/01/21 업데이트 될 때 routeRecordDate도 업데이트
     @Transactional
-    public ClimbingRecordSimpleInfo updateClimbingRecord(Long id,
+    public ClimbingRecordSimpleInfo updateClimbingRecord(User user, Long id,
         UpdateClimbingRecordDto requestDto) {
 
         ClimbingRecord climbingRecord = climbingRecordRepository.findById(id)
             .orElseThrow(() -> new GeneralException(ErrorStatus._CLIMBING_RECORD_NOT_FOUND));
+
+        if(!user.getId().equals(climbingRecord.getUser().getId())){
+            throw new GeneralException(ErrorStatus._INVALID_MEMBER);
+        }
 
         LocalDate oldDate = climbingRecord.getClimbingDate();
         LocalTime oldTime = climbingRecord.getClimbingTime();
@@ -149,9 +157,14 @@ public class ClimbingRecordService {
     }
 
     @Transactional
-    public ResponseEntity<String> deleteClimbingRecord(Long id) {
+    public ResponseEntity<String> deleteClimbingRecord(User user, Long id) {
+
         ClimbingRecord climbingRecord = climbingRecordRepository.findById(id)
             .orElseThrow(() -> new GeneralException(ErrorStatus._CLIMBING_RECORD_NOT_FOUND));
+
+        if(!user.getId().equals(climbingRecord.getUser().getId())){
+            throw new GeneralException(ErrorStatus._INVALID_MEMBER);
+        }
 
         climbingRecordRepository.deleteById(id);
 
@@ -160,12 +173,12 @@ public class ClimbingRecordService {
         return ResponseEntity.ok("클라이밍기록이 삭제되었습니다.");
     }
 
-    public ClimbingRecordStatisticsInfo getClimbingRecordStatistics(int year, int month) {
+    public ClimbingRecordStatisticsInfo getClimbingRecordStatistics(User user, int year, int month) {
 
         LocalDate startDate = LocalDate.of(year, month, START_DAY_OF_MONTH);
         LocalDate endDate = YearMonth.of(year, month).atEndOfMonth();
 
-        Tuple crTuple = climbingRecordRepository.getStatisticsInfoBetween(startDate, endDate);
+        Tuple crTuple = climbingRecordRepository.getStatisticsInfoBetween(user, startDate, endDate);
 
         if (crTuple.get("totalTime") == null) {
             throw new GeneralException(ErrorStatus._EMPTY_CLIMBING_RECORD);
